# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VideoGames API",
        "description": "API REST para la gestión de videojuegos, reseñas, compras y usuarios. Proyecto Laravel 12 con MySQL.",
        "contact": {
            "name": "Equipo VideoGames API",
            "email": "soporte@videogamesapi.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/platforms": {
            "get": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Listar todas las plataformas",
                "operationId": "b79ea4224ff39ae923cc8756fa1bf671",
                "responses": {
                    "200": {
                        "description": "Lista de plataformas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Platform"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Crear una nueva plataforma",
                "operationId": "057364b51fcd5d4e282c22cbc0e92eaf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Platform"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plataforma creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/platforms/{id}": {
            "get": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Obtener detalles de una plataforma",
                "operationId": "3deec4f1d4d2c00065c8ec17fd57a93b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plataforma encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Actualizar una plataforma",
                "operationId": "4d11c4ec68257f86e884985341d6f516",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Platform"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plataforma actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Eliminar una plataforma",
                "operationId": "4e4d8de139b615a76288e770f777e4c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/publishers": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Listar todos los editores",
                "operationId": "9a0fdd4383afd0c4b7d0c0920a9e2d09",
                "responses": {
                    "200": {
                        "description": "Lista de editores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Publisher"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Crear un nuevo editor",
                "operationId": "c3d281ee2ed8ec68f724ca9ffb658b8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Publisher"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Editor creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publisher"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/publishers/{id}": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Obtener un editor",
                "operationId": "ae45ae012b4a2309ec4fb29672f3db1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Editor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publisher"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Actualizar un editor",
                "operationId": "7271f0419620ecbd4d293ba3943bf80d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Publisher"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Editor actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publisher"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Eliminar un editor",
                "operationId": "a7b6cb84d66d2637167a22db043213c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Platform": {
                "title": "Platform",
                "description": "Plataforma de videojuego",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "PlayStation 5"
                    },
                    "manufacturer": {
                        "type": "string",
                        "example": "Sony"
                    }
                },
                "type": "object"
            },
...
            "Publisher": {
                "title": "Publisher",
                "description": "Editor o distribuidor de videojuegos",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sony Interactive Entertainment"
                    },
                    "country": {
                        "type": "string",
                        "example": "USA"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}